@@ -28,62 +28,61 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "^17.0.2",
-    "@angular/common": "^17.0.2",
-    "@angular/compiler": "^17.0.2",
-    "@angular/core": "^17.0.2",
-    "@angular/forms": "^17.0.2",
-    "@angular/platform-browser": "^17.0.2",
-    "@angular/platform-browser-dynamic": "^17.0.2",
-    "@angular/router": "^17.0.2",
-    "bugsplat": "^8.0.0",
-    "core-js": "^2.6.3",
-    "rxjs": "^6.6.3",
-    "tslib": "^2.3.0",
-    "zone.js": "^0.14.2"
+    "@angular/animations": "^18.0.6",
+    "@angular/common": "^18.0.6",
+    "@angular/compiler": "^18.0.6",
+    "@angular/core": "^18.0.6",
+    "@angular/forms": "^18.0.6",
+    "@angular/platform-browser": "^18.0.6",
+    "@angular/platform-browser-dynamic": "^18.0.6",
+    "@angular/router": "^18.0.6",
+    "bugsplat": "^8.0.1",
+    "core-js": "^3.37.1",
+    "rxjs": "^7.8.1",
+    "tslib": "^2.6.3",
+    "zone.js": "^0.14.7"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^17.0.0",
-    "@angular-eslint/builder": "17.0.1",
-    "@angular-eslint/eslint-plugin": "17.0.1",
-    "@angular-eslint/eslint-plugin-template": "17.0.1",
-    "@angular-eslint/schematics": "17.0.1",
-    "@angular-eslint/template-parser": "17.0.1",
-    "@angular/cli": "^17.0.0",
-    "@angular/compiler-cli": "^17.0.2",
-    "@angular/language-service": "^17.0.2",
-    "@bugsplat/symbol-upload": "^8.0.6",
-    "@commitlint/cli": "^17.0.3",
-    "@commitlint/config-conventional": "^17.0.3",
-    "@semantic-release/changelog": "^6.0.1",
+    "@angular-devkit/build-angular": "^18.0.7",
+    "@angular-eslint/builder": "18.1.0",
+    "@angular-eslint/eslint-plugin": "18.1.0",
+    "@angular-eslint/eslint-plugin-template": "18.1.0",
+    "@angular-eslint/schematics": "18.1.0",
+    "@angular-eslint/template-parser": "18.1.0",
+    "@angular/cli": "^18.0.7",
+    "@angular/compiler-cli": "^18.0.6",
+    "@angular/language-service": "^18.0.6",
+    "@bugsplat/symbol-upload": "^9.1.0",
+    "@commitlint/cli": "^19.3.0",
+    "@commitlint/config-conventional": "^19.2.2",
+    "@semantic-release/changelog": "^6.0.3",
     "@semantic-release/git": "^10.0.1",
-    "@types/jasmine": "~3.3.8",
-    "@types/jasminewd2": "~2.0.6",
-    "@types/node": "^12.20.55",
-    "@typescript-eslint/eslint-plugin": "^6.10.0",
-    "@typescript-eslint/parser": "^6.10.0",
+    "@types/jasmine": "^5.1.4",
+    "@types/jasminewd2": "^2.0.13",
+    "@types/node": "^20.14.9",
+    "@typescript-eslint/eslint-plugin": "^7.15.0",
+    "@typescript-eslint/parser": "^7.15.0",
     "copyfiles": "^2.4.1",
-    "dotenv": "^10.0.0",
-    "eslint": "^8.53.0",
-    "eslint-plugin-import": "^2.26.0",
-    "eslint-plugin-jsdoc": "^39.3.3",
+    "dotenv": "^16.4.5",
+    "eslint": "^8.57.0",
+    "eslint-plugin-import": "^2.29.1",
+    "eslint-plugin-jsdoc": "^48.5.2",
     "eslint-plugin-prefer-arrow": "^1.2.3",
     "http-server": "^14.1.1",
-    "husky": "^8.0.1",
+    "husky": "^9.0.11",
     "jasmine-auto-spies": "^8.0.1",
-    "jasmine-core": "^3.99.1",
-    "jasmine-spec-reporter": "~5.0.0",
-    "karma": "~6.3.16",
-    "karma-chrome-launcher": "~3.1.0",
-    "karma-coverage-istanbul-reporter": "~3.0.2",
-    "karma-jasmine": "~4.0.0",
-    "karma-jasmine-html-reporter": "^1.5.0",
+    "jasmine-core": "^5.1.2",
+    "jasmine-spec-reporter": "~7.0.0",
+    "karma": "~6.4.3",
+    "karma-chrome-launcher": "~3.2.0",
+    "karma-coverage-istanbul-reporter": "~3.0.3",
+    "karma-jasmine": "~5.1.0",
+    "karma-jasmine-html-reporter": "^2.1.0",
     "karma-teamcity-reporter": "^1.1.0",
-    "ng-packagr": "^17.0.0",
-    "semantic-release": "^21.0.7",
-    "semantic-release-cli": "^5.2.1",
-    "ts-node": "^10.9.1",
-    "typescript": "~5.2.2"
+    "ng-packagr": "^18.0.0",
+    "semantic-release": "^24.0.0",
+    "ts-node": "^10.9.2",
+    "typescript": "^5.4.5"
   },
   "repository": {
     "type": "git",
