@@ -1,4 +1,4 @@
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { ModuleWithProviders, NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { BugSplat as BugSplatJs } from 'bugsplat';
@@ -6,12 +6,7 @@ import { BugSplat } from './bugsplat';
 import { BugSplatErrorHandler } from './bugsplat-error-handler';
 import { BugSplatSettings } from './bugsplat-settings';
 
-@NgModule({
-    imports: [
-        BrowserModule,
-        HttpClientModule
-    ]
-})
+@NgModule({ imports: [BrowserModule], providers: [provideHttpClient(withInterceptorsFromDi())] })
 export class BugSplatModule { 
     static initializeApp(
         settings: BugSplatSettings
