@@ -1,4 +1,4 @@
-import { HttpClientModule } from '@angular/common/http';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 import { ErrorHandler, NgModule } from '@angular/core';
 import { BrowserModule } from '@angular/platform-browser';
 import { BugSplatLogger, BugSplatLogLevel, BugSplatModule } from 'bugsplat-ng';
@@ -20,29 +20,20 @@ import { MyAngularErrorHandler } from './my-angular-error-handler';
 // })
 // Comment this to use the simple error handler
 
-@NgModule({
-  declarations: [
-    AppComponent
-  ],
-  imports: [
-    BrowserModule,
-    HttpClientModule,
-    BugSplatModule.initializeApp(environment.bugsplat)
-  ],
-  providers: [
-    {
-      provide: ErrorHandler,
-      useClass: MyAngularErrorHandler
-    },
-    {
-      provide: BugSplatLogger,
-      useValue: new BugSplatLogger(
-        BugSplatLogLevel.info,
-        console
-      )
-    }
-  ],
-  bootstrap: [AppComponent]
-})
+@NgModule({ declarations: [
+        AppComponent
+    ],
+    bootstrap: [AppComponent], imports: [BrowserModule,
+        BugSplatModule.initializeApp(environment.bugsplat)], providers: [
+        {
+            provide: ErrorHandler,
+            useClass: MyAngularErrorHandler
+        },
+        {
+            provide: BugSplatLogger,
+            useValue: new BugSplatLogger(BugSplatLogLevel.info, console)
+        },
+        provideHttpClient(withInterceptorsFromDi())
+    ] })
 
 export class AppModule { }
