@@ -1,12 +1,14 @@
-import { HttpClientTestingModule } from '@angular/common/http/testing';
+import { provideHttpClientTesting } from '@angular/common/http/testing';
 import { TestBed } from '@angular/core/testing';
 import { BugSplatResponseData, SuccessResponse } from '../lib/bugsplat-response-data';
+import { provideHttpClient, withInterceptorsFromDi } from '@angular/common/http';
 
 describe('BugSplatResponseData', () => {
 
     beforeEach(() => TestBed.configureTestingModule({
-        imports: [HttpClientTestingModule],
-    }));
+    imports: [],
+    providers: [provideHttpClient(withInterceptorsFromDi()), provideHttpClientTesting()]
+}));
 
     it('should return BugSplatResponseData when createFromSuccessResponseObject is called with valid parameters', () => {
         const response = {} as SuccessResponse;
